Did not highlight changes to double precision registers correctly. Unnecessary refresh of register display.
git-svn-id: http://svn.code.sf.net/p/spimsimulator/code@689 663c0655-963b-4438-a250-40b5657b8e18

+++ QtSpim/spimview.h
@@ -189,7 +189,7 @@ Q_OBJECT
     QString formatDFPRegisters();
     QString formatDFPRegister(int regNum, double value, bool changed);
 
-    float oldFPR_D[FPR_LENGTH];
+    double oldFPR_D[FPR_LENGTH];
 
 
     // Common register methods
